Performance fix
Turns out allocating a mutableMap was tanking the performance so now I just draw each pixel as soon as I get its color value. This has sped it up by nearly 10x which is awesome. I also updated the keyboard callback so it only refreshes the view on a keypress that modifies the view. Kept redrawing when I wanted to screenshot and that was driving me crazy!

Next on the list is to maybe multithread it and get a better color pallet going. At higher zoom levels the colors just get less interesting right now.

Quad precision is also high on the want list as there is a definite limit when zooming in. The fractals stop diverging and become smooth curves.