Zoom to center + More fine tuning
Previously when zooming it would zoom in on the upper-left corner which made it difficult to zoom in without having to repan the view. Now it zooms to the center of the view, making life so much easier.

I also added a couple new keyboard shortcuts for adjusting the quality of the visualization on the fly. The +/- keys control the max number of iterations to run when testing escape velocity. This has a massive effect on quality, moreso than point size. Point size can also be adjusted using the keys directly below, the [ and ] keys. The window title now update with the zoom level, current locations, etc. Still need to multithread it and get a better color pallet working.