ease of use update
Added some niceties to make the app a little easier to work with and require less recompilation to change things.

Added the ability to adjust the color mode during viewing by using the A and S keys. A toggles through the 3 basic color modes while S toggles through trig-based modes. Next color modes to be added will be table-based, both discrete and gradiated.

I also added the ability to set the current coordinates by clicking. Using the arrow keys to navigate was nice but was too imprecise.

Also decided to lock the window aspect ratio to 3:2 for optimum viewing. This means now only the window width is configurable.