GLSL update
Updated to use GLSL instead of running on the CPU. The speedup is incredible. Because of some limitations with fp64 functions in GLSL I had to remove the trig-based color modes. Also the point size is now set to 1 and can't be changed. Might see if I can change this in the future to get the old behaior back for both the colors and trig modes. Will be a good time to get the table-based color mode working though.