Performance and usability improvements
Resolved a CPU usage issue, dropping usage in worst case from 50% to 25% after switching from swapBuffers() to glFlush(). The other 25% appears to be related to the glUniform*() calls but I'm unsure how I can speed those up.

Fixed zooming to be more linear when zooming out. Previously, zooming out didn't actually change the same as the previous zoom inward did. This messed up the zoom level number completely. I tried to think of a nice solution in maths but today isn't my day. I ended up just implementing a zoom stack and on unzoom we just pop the last zoom level and set to that. Works fine.

I added in a 32-bit version of the fragment shader just to compare the quality of varying zoom levels. The same difference is expected when I get >fp64 implemented. You can switch current FP modes using the 1 and 2 keys on the numpad.

Finally I added in 3 location bookmarks for convenience. Using the 1, 2, and 3 keys on the keyboard will move the view to those locations.