Failed speedup
I'm a bit annoyed with this code because last night I swear it was on parity in terms of performance (about 73ms to draw worse case) but at the moment this code takes 100ms extra to draw over that. Posting the code here because it took some time to figure out how to use buffers properly.

Next attempt will be to revert to normal uniform calls but to collect all the int and double uniforms into just two vec4's instead. This should drop the numer of uniform calls (compared to the previous code that ran at 73ms) from 8 to 2. If the uniform calls are truly the performance issue this will definitely improve things.

Also, forgot to use glClear this whole time. Thats kinda weird. Its not causing the new performance penalty however (I wish it were that simple).