Opengl 3.3 compatibility + bugfixes
Wouldn't run on a laptop I had so I rewrote parts to be more compatible with older hardware. It now can fall back to OpenGL v3.3 if the hardware doesn't support fp64 extensions. Because version 330 of GLSL doesn't support the offset layout option in uniforms we have to use default offsets. This means when we switch modes (if the hw supports both) we have to redo the indices and offsets on the kotlin side. Works on the laptop now but I havent tested it on the better hardware.

Fixed a bug where even if you weren't saving images of a zoom sequence it would still create the directory for it.