bugfix/cleanup
Didn't quite work on the main dev box because of some reference issues (differences between lwjgl versions I suspect) but that was quickly fixed. One other minor bug prevented using fp64 mode because we needed to only run a glUseProgram() before we actually used the shader program. We were always running the command on the 32-bit shader and never ran it on the 64-bit shader. Now they run appropriately. Got rid of the uniform layout stuff in the 64-bit shader since it was only instructional before and now its completely useless.