Bifurcation Diagram
Had always wanted to add this in but never got it done. Pressing B on the keyboard will bring up the bifurcation diagram which shows the different values obtained from the mandelbrot set equation when iterated over the horizontal axis. You can visually see how the equation convergest to a single values on the right but then diverges to an oscilation between 2 different values, then 4, etc. These points are gathered after the equation has run for a number of iterations first. You can adjust the iteration count by changing j (as well as the offset to j defining when we start displaying points. see ln#389). I suggest lowering the values to 10 and 5 and seeing what you get. After a certain iteration count the results don't get appreciably better so we stop at 200. Would be interesting to be able to zoom into the far left side.

I'm hoping to clean this up into its own shader at some point.